CORE-17955 - Update
