+++ components/ledger/ledger-utxo-token-cache/src/main/kotlin/net/corda/ledger/utxo/token/cache/factories/TokenCacheEventProcessorFactoryImpl.kt
@@ -1,8 +1,5 @@
 package net.corda.ledger.utxo.token.cache.factories
 
-import net.corda.data.ledger.utxo.token.selection.event.TokenPoolCacheEvent
-import net.corda.data.ledger.utxo.token.selection.key.TokenPoolCacheKey
-import net.corda.data.ledger.utxo.token.selection.state.TokenPoolCacheState
 import net.corda.db.connection.manager.DbConnectionManager
 import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
 import net.corda.ledger.utxo.token.cache.converters.EntityConverter
@@ -17,18 +14,9 @@ import net.corda.ledger.utxo.token.cache.handlers.TokenForceClaimReleaseEventHan
 import net.corda.ledger.utxo.token.cache.handlers.TokenLedgerChangeEventHandler
 import net.corda.ledger.utxo.token.cache.queries.impl.SqlQueryProviderTokens
 import net.corda.ledger.utxo.token.cache.repositories.impl.UtxoTokenRepositoryImpl
-import net.corda.ledger.utxo.token.cache.services.ClaimStateStoreCacheImpl
-import net.corda.ledger.utxo.token.cache.services.ClaimStateStoreFactoryImpl
-import net.corda.ledger.utxo.token.cache.services.ServiceConfiguration
-import net.corda.ledger.utxo.token.cache.services.SimpleTokenFilterStrategy
-import net.corda.ledger.utxo.token.cache.services.TokenCacheEventProcessor
-import net.corda.ledger.utxo.token.cache.services.TokenPoolCacheStateSerialization
-import net.corda.ledger.utxo.token.cache.services.TokenSelectionDelegatedProcessor
-import net.corda.ledger.utxo.token.cache.services.TokenSelectionDelegatedProcessorImpl
-import net.corda.ledger.utxo.token.cache.services.TokenSelectionMetricsImpl
+import net.corda.ledger.utxo.token.cache.services.*
 import net.corda.ledger.utxo.token.cache.services.internal.AvailableTokenServiceImpl
 import net.corda.libs.statemanager.api.StateManager
-import net.corda.messaging.api.processor.StateAndEventProcessor
 import net.corda.orm.JpaEntitiesRegistry
 import net.corda.utilities.time.Clock
 import net.corda.utilities.time.UTCClock
@@ -49,12 +37,15 @@ class TokenCacheEventProcessorFactoryImpl constructor(
 
     private val tokenPoolCache = TokenPoolCacheImpl()
 
-    override fun create(): StateAndEventProcessor<TokenPoolCacheKey, TokenPoolCacheState, TokenPoolCacheEvent> {
+    override fun createTokenSelectionSyncRPCProcessor(
+        stateManager: StateManager
+    ): TokenSelectionSyncRPCProcessor {
+        val claimStateStoreFactory = ClaimStateStoreFactoryImpl(stateManager, serialization, tokenPoolCache, clock)
+        val tokenSelectionMetrics = TokenSelectionMetricsImpl(UTCClock())
         val recordFactory = RecordFactoryImpl(externalEventResponseFactory)
         val tokenFilterStrategy = SimpleTokenFilterStrategy()
         val sqlQueryProvider = SqlQueryProviderTokens()
         val utxoTokenRepository = UtxoTokenRepositoryImpl(sqlQueryProvider)
-        val tokenSelectionMetrics = TokenSelectionMetricsImpl(clock)
         val availableTokenService = AvailableTokenServiceImpl(
             virtualNodeInfoService,
             dbConnectionManager,
@@ -77,25 +68,11 @@ class TokenCacheEventProcessorFactoryImpl constructor(
             createHandler(TokenLedgerChangeEventHandler()),
             createHandler(TokenBalanceQueryEventHandler(recordFactory, availableTokenService)),
         )
-        return TokenCacheEventProcessor(
+        return TokenSelectionSyncRPCProcessor(
             eventConverter,
             entityConverter,
-            tokenPoolCache,
             eventHandlerMap,
-            externalEventResponseFactory
-        )
-    }
-
-    override fun createDelegatedProcessor(
-        stateManager: StateManager,
-        processor: StateAndEventProcessor<TokenPoolCacheKey, TokenPoolCacheState, TokenPoolCacheEvent>
-    ): TokenSelectionDelegatedProcessor {
-        val claimStateStoreFactory = ClaimStateStoreFactoryImpl(stateManager, serialization, tokenPoolCache, clock)
-        val tokenSelectionMetrics = TokenSelectionMetricsImpl(UTCClock())
-        return TokenSelectionDelegatedProcessorImpl(
-            eventConverter,
-            entityConverter,
-            create(),
+            tokenPoolCache,
             ClaimStateStoreCacheImpl(stateManager, serialization, claimStateStoreFactory, clock),
             externalEventResponseFactory,
             tokenSelectionMetrics
